requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
gitpython>=3.1.0
